% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_vectorfield.R
\name{add_interp_grid}
\alias{add_interp_grid}
\title{Add a grid to a \code{vectorfield} object to enable linear interpolation}
\usage{
add_interp_grid(vf, lims = vf$lims, n = vf$n)
}
\arguments{
\item{vf}{A \code{vectorfield} object estimated by \code{\link[=fit_2d_vf]{fit_2d_vf()}}.}

\item{lims}{The limits of the range for the vector field estimation as \verb{c(<xl>, <xu>, <yl>, <yu>)}. If missing, the range of the data extended by 10\% for both sides will be used.}

\item{n}{The number of equally spaced points in each axis, at which the vectors are to be estimated.}
}
\value{
A \code{vectorfield} project with an \code{interp_grid} field.
}
\description{
Add a grid to a \code{vectorfield} object to enable linear interpolation
}
