% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{autoplot.1d_vectorfield}
\alias{autoplot.1d_vectorfield}
\title{Autoplot a 1D vector field}
\usage{
\method{autoplot}{`1d_vectorfield`}(
  object,
  show_estimated_vector = TRUE,
  estimated_vector_options = list(),
  show_point = TRUE,
  point_options = list(size = 0.5),
  show_original_vector = FALSE,
  original_vector_options = list(),
  ...
)
}
\arguments{
\item{object}{A \verb{1d_vectorfield} object estimated by \code{\link[=fit_1d_vf]{fit_1d_vf()}}.}

\item{show_estimated_vector}{Show the estimated drift curve? \code{TRUE} by
default.}

\item{estimated_vector_options}{A list passing customized parameters to
\code{\link[ggplot2:geom_line]{ggplot2::geom_line()}} for the estimated drift curve.}

\item{show_point}{Show the original data points? \code{TRUE} by default.}

\item{point_options}{A list passing customized parameters to
\code{\link[ggplot2:geom_point]{ggplot2::geom_point()}} for the original data points.}

\item{show_original_vector}{Show the original drift vectors? \code{FALSE} by
default.}

\item{original_vector_options}{A list passing customized parameters to
\code{\link[ggplot2:geom_point]{ggplot2::geom_point()}} for the original drift vectors.}

\item{...}{Not in use.}
}
\value{
A \code{ggplot2} plot.
}
\description{
Plot a 1D vector field estimated by \code{\link[=fit_1d_vf]{fit_1d_vf()}}. Powered by
\code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}.
}
