% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_api.R
\name{autoplot.2d_stream}
\alias{autoplot.2d_stream}
\title{Autoplot a two-dimensional stream function}
\usage{
\method{autoplot}{`2d_stream`}(object, contour = TRUE, ...)
}
\arguments{
\item{object}{A \verb{2d_stream} object returned by \code{\link[=make_2d_stream]{make_2d_stream()}}.}

\item{contour}{Logical indicating whether to overlay contour lines.}

\item{...}{Additional arguments, currently unused.}
}
\value{
A ggplot object.
}
\description{
Autoplot a two-dimensional stream function
}
