% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{autoplot.cv_vectorfield}
\alias{autoplot.cv_vectorfield}
\title{Autoplot 2D vector-field cross-validation results}
\usage{
\method{autoplot}{cv_vectorfield}(object, ...)
}
\arguments{
\item{object}{An object of class 'cv_vectorfield' returned by cv_fit_2d_vf.}

\item{...}{Additional arguments (not used).}
}
\value{
A ggplot object.
}
\description{
Draw the cross-validation error over the candidate bandwidth values and
mark the selected bandwidth.
}
