% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscape_clustering.R
\name{autoplot.landscape_cluster_evaluation}
\alias{autoplot.landscape_cluster_evaluation}
\title{Autoplot landscape-cluster evaluation}
\usage{
\method{autoplot}{landscape_cluster_evaluation}(object, ...)
}
\arguments{
\item{object}{A \code{landscape_cluster_evaluation} object.}

\item{...}{Additional arguments, currently unused.}
}
\value{
A ggplot elbow plot.
}
\description{
Autoplot landscape-cluster evaluation
}
