% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscape_clustering.R
\name{autoplot.landscape_clusters}
\alias{autoplot.landscape_clusters}
\title{Autoplot landscape-clustering centers}
\usage{
\method{autoplot}{landscape_clusters}(object, ...)
}
\arguments{
\item{object}{A \code{landscape_clusters} object returned by
\code{\link[=cluster_landscapes]{cluster_landscapes()}}.}

\item{...}{Additional arguments, currently unused.}
}
\value{
A faceted ggplot of cluster-mean landscapes.
}
\description{
Autoplot landscape-clustering centers
}
