% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream_clustering.R
\name{autoplot.landscape_stream_cluster_evaluation}
\alias{autoplot.landscape_stream_cluster_evaluation}
\title{Autoplot joint landscape-stream cluster evaluation}
\usage{
\method{autoplot}{landscape_stream_cluster_evaluation}(object, ...)
}
\arguments{
\item{object}{A \code{landscape_stream_cluster_evaluation} object.}

\item{...}{Additional arguments, currently unused.}
}
\value{
A ggplot elbow plot.
}
\description{
Autoplot joint landscape-stream cluster evaluation
}
