% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream_clustering.R
\name{autoplot.landscape_stream_clusters}
\alias{autoplot.landscape_stream_clusters}
\title{Autoplot joint landscape-stream cluster centers}
\usage{
\method{autoplot}{landscape_stream_clusters}(object, type = c("landscape", "stream"), ...)
}
\arguments{
\item{object}{A \code{landscape_stream_clusters} object.}

\item{type}{Which center representation to draw, \code{"landscape"} or
\code{"stream"}.}

\item{...}{Additional arguments passed to the selected center autoplot.}
}
\value{
A faceted ggplot of landscape or stream centers.
}
\description{
Autoplot joint landscape-stream cluster centers
}
