% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream_clustering.R
\name{autoplot.stream_cluster_evaluation}
\alias{autoplot.stream_cluster_evaluation}
\title{Autoplot stream-cluster evaluation}
\usage{
\method{autoplot}{stream_cluster_evaluation}(object, ...)
}
\arguments{
\item{object}{A \code{stream_cluster_evaluation} object.}

\item{...}{Additional arguments, currently unused.}
}
\value{
A ggplot elbow plot.
}
\description{
Autoplot stream-cluster evaluation
}
