% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream_clustering.R
\name{autoplot.stream_clusters}
\alias{autoplot.stream_clusters}
\title{Autoplot stream-clustering centers}
\usage{
\method{autoplot}{stream_clusters}(object, contour = TRUE, ...)
}
\arguments{
\item{object}{A \code{stream_clusters} object returned by \code{\link[=cluster_streams]{cluster_streams()}}.}

\item{contour}{Logical indicating whether to overlay contour lines.}

\item{...}{Additional arguments, currently unused.}
}
\value{
A faceted ggplot of cluster-mean stream functions.
}
\description{
Autoplot stream-clustering centers
}
