% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapping_1d.R
\name{bootstrap_1d_ld}
\alias{bootstrap_1d_ld}
\title{Bootstrap 1D landscapes}
\usage{
bootstrap_1d_ld(boot_vf, ...)
}
\arguments{
\item{boot_vf}{A \code{bootstrap_1d_vf} object.}

\item{...}{Additional arguments passed to \code{\link[=make_1d_ld]{make_1d_ld()}}.}
}
\value{
A \code{bootstrap_1d_ld} object.
}
\description{
Bootstrap 1D landscapes
}
