% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapping.R
\name{bootstrap_2d_vf}
\alias{bootstrap_2d_vf}
\title{Performs bootstrap resampling to estimate uncertainty in 2D vector field estimation.}
\usage{
bootstrap_2d_vf(
  vf,
  block_length = NULL,
  n_boot = 200,
  seed = 1614,
  add_linear_interp = TRUE,
  ...
)
}
\arguments{
\item{vf}{A \code{vectorfield} or \code{cv_vectorfield} object representing the fitted vector field.}

\item{block_length}{Length of each block for MBB. If NULL, it will be set to n^(1/3).}

\item{n_boot}{Number of bootstrap samples to generate (default 200).}

\item{seed}{Random seed for reproducibility (default 1614).}

\item{add_linear_interp}{Logical indicating whether to add linear interpolation in predictions.}

\item{...}{Additional arguments passed to the vector field fitting function.}
}
\value{
An object of class \code{bootstrap_2d_vf} containing:
- \code{bootstrap_models}: A list of fitted vector field models from each bootstrap sample.
- \code{original_vf}: The original fitted vector field model.
- \code{block_length}: The block length used for MBB.
- \code{n_boot}: The number of bootstrap samples.
}
\description{
Moving block bootstrap (MBB) is used to account for temporal dependencies in the data.
}
