% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_flow.R
\name{calculate_probability_flow}
\alias{calculate_probability_flow}
\title{Calculates Probability Flow (J) vectors at a subset of grid points}
\usage{
calculate_probability_flow(
  rho,
  drift_func,
  diffusion_func,
  x_range = c(0, 1),
  y_range = c(0, 1),
  n_flow = 20,
  devided_by_rho = FALSE
)
}
\arguments{
\item{rho}{The steady-state density matrix (n_grid x n_grid).}

\item{drift_func}{Function returning c(Ax, Ay).}

\item{diffusion_func}{Function returning 2x2 matrix.}

\item{x_range}{Vector c(xmin, xmax).}

\item{y_range}{Vector c(ymin, ymax).}

\item{n_flow}{Number of points along one dimension to sample (default 20).}
}
\value{
Data frame with columns: x, y, Jx, Jy.
}
\description{
Calculates Probability Flow (J) vectors at a subset of grid points
}
