% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseVFC.R
\name{con_K}
\alias{con_K}
\title{Construct the kernel K}
\usage{
con_K(x, y, beta)
}
\arguments{
\item{x, y, beta}{The variable and parameter values for the funcion
described above.}
}
\description{
\deqn{K[i, j] = k(x[i,], y[j,]) = exp(-beta*||x[i,]-y[j,]||^2)}
}
\seealso{
\code{\link[=SparseVFC]{SparseVFC()}}
}
