% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_vectorfield_1d.R
\name{cv_fit_1d_vf}
\alias{cv_fit_1d_vf}
\title{10-fold cross-validation for 1D vector fields}
\usage{
cv_fit_1d_vf(
  data,
  x,
  dayvar = NULL,
  beepvar = NULL,
  h_values = exp(seq(log(0.01), log(2), length.out = 20)),
  k = 10,
  ...
)
}
\arguments{
\item{data}{The data set used for estimating the vector field.}

\item{x}{Character naming the state variable.}

\item{dayvar}{Optional character scalar naming the day variable.}

\item{beepvar}{Optional character scalar naming the within-day assessment
order variable.}

\item{h_values}{Numeric vector of candidate bandwidths.}

\item{k}{Number of folds.}

\item{...}{Additional arguments passed to \code{\link[=MVKE]{MVKE()}} or
\code{\link[SparseVFC:SparseVFC]{SparseVFC::SparseVFC()}}.}
}
\value{
A \code{cv_vectorfield} object.
}
\description{
10-fold cross-validation for 1D vector fields
}
