% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_vectorfield.R
\name{fields_bilinear}
\alias{fields_bilinear}
\title{To make fields::interp.surface accept the parameters of akima::bilinear and behave like it}
\usage{
fields_bilinear(x, y, z, x0, y0)
}
\arguments{
\item{x}{
    a vector containing the \code{x} coordinates of the rectangular data grid.
  }

\item{y}{
    a vector containing the \code{y} coordinates of the rectangular data grid.
  }

\item{z}{
    a matrix containing the \code{z[i,j]} data values for the grid points (\code{x[i]},\code{y[j]}).
  }

\item{x0}{
    vector of \code{x} coordinates used to interpolate at.
  }

\item{y0}{
    vector of \code{y} coordinates used to interpolate at.
  }
}
\description{
To make fields::interp.surface accept the parameters of akima::bilinear and behave like it
}
