% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorfield_eq.R
\name{find_eqs}
\alias{find_eqs}
\title{Find equilibrium points for a vector field}
\usage{
find_eqs(vf, starts, jacobian_params = list(), ...)
}
\arguments{
\item{vf}{A \code{vectorfield} object estimated by \code{\link[=fit_2d_vf]{fit_2d_vf()}}.}

\item{starts}{A vector indicating the starting value for solving the equilibrium point, or a list of vectors providing multiple starting values together.}

\item{jacobian_params}{Parameters passed to \code{\link[numDeriv:jacobian]{numDeriv::jacobian()}}.}

\item{...}{Parameters passed to \code{\link[rootSolve:multiroot]{rootSolve::multiroot()}}.}
}
\value{
A list of equilibrium points and their details. Use \code{print.vectorfield_eqs()} to inspect it.
}
\description{
Find equilibrium points for a vector field
}
