% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_path.R
\name{find_gmam_path}
\alias{find_gmam_path}
\title{Calculate the Minimum Action Path using gMAM}
\usage{
find_gmam_path(
  vf,
  linear_interp = TRUE,
  start,
  end,
  n_beads = 10,
  max_iter = 10,
  progress = interactive()
)
}
\arguments{
\item{vf}{A \code{vectorfield} or \code{cv_vectorfield} object.}

\item{linear_interp}{Logical indicating whether to use linear interpolation in predictions.}

\item{start}{A numeric vector c(x, y) for the starting point.}

\item{end}{A numeric vector c(x, y) for the destination point.}

\item{n_beads}{Number of points to represent the path (default 50).}

\item{max_iter}{Number of optimization-reparameterization cycles.}
}
\value{
A data frame containing the x and y coordinates of the optimal path.
}
\description{
Calculate the Minimum Action Path using gMAM
}
