% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorfield_simulation.R
\name{fit_2d_vfld}
\alias{fit_2d_vfld}
\title{Estimate a 2D potential landscape from a vector field}
\usage{
fit_2d_vfld(
  vf,
  method = "simlandr",
  .sim_vf_options = sim_vf_options(vf),
  .simlandr_options = simlandr_options(vf)
)
}
\arguments{
\item{vf}{A \code{vectorfield} object estimated by \code{\link[=fit_2d_vf]{fit_2d_vf()}}.]}

\item{method}{The method used for landscape construction. Currently "simlandr" is the only possible choice.}

\item{.sim_vf_options}{Options controlling the vector field simulation. Should be generated by \code{\link[=sim_vf_options]{sim_vf_options()}}.}

\item{.simlandr_options}{Options controlling the landscape construction. Should be generated by \code{\link[=simlandr_options]{simlandr_options()}}.}
}
\description{
For \code{method = "simlandr"}, the landscape is constructed based on the generalized potential landscape by Wang et al (2008), implemented by the \code{simlandr} package. This function is a wraper of \code{\link[=sim_vf]{sim_vf()}} and \code{\link[simlandr:make_3d_static]{simlandr::make_3d_static()}}. Use those two functions seperately for more customization.
}
