% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorfield_landscape.R
\name{fit_vf_2d}
\alias{fit_vf_2d}
\title{Estimate a 2D vector field}
\usage{
fit_vf_2d(
  d,
  x,
  y,
  x_start,
  x_end,
  x_by,
  y_start,
  y_end,
  y_by,
  vector_position = "middle",
  na_action = "omit_data_points",
  method = "VFC",
  ...
)
}
\arguments{
\item{d}{The data set used for estimating the vector field.
Should be a data frame or a matrix.}

\item{x, y}{Characters to indicate the name of the two variables.}

\item{x_start, x_end, x_by, y_start, y_end, y_by}{Specify the sample grid for the output: the starting, ending, and increment of the sequences for both axes.}

\item{vector_position}{One of "start", "middle", or "end". If "start",}

\item{na_action}{One of "omit_data_points" or "omit_vectors". If using "omit_data_points", then the vectors will be retained even if there are \code{NA} points between them. If using "omit_vectors", then the vectors will be omitted if either of its points is \code{NA}.}

\item{method}{One of "VFC" or "MVKE".}

\item{...}{Other parameters to be passed to \code{\link[SparseVFC:SparseVFC]{SparseVFC::SparseVFC()}} or \code{\link[=MVKE]{MVKE()}}.}
}
\value{
A \code{vectorfield} object.
}
\description{
Estimate a 2D vector field from intensive longitudinal data. Two methods can be used: Sparse Vector Field Consensus (SparseVFC, using \code{\link[SparseVFC:SparseVFC]{SparseVFC::SparseVFC()}}), or Multivariate Vector Field Kernel Estimator (MVKE, using \code{\link[=MVKE]{MVKE()}}).
}
\seealso{
\code{\link[=plot.vectorfield]{plot.vectorfield()}}
}
