% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorfield_landscape.R
\name{fit_vfld_2d}
\alias{fit_vfld_2d}
\title{Estimate a 2D potential landscape from a vector field}
\usage{
fit_vfld_2d(
  vf,
  x_start = vf$x_start,
  x_end = vf$x_end,
  x_by = vf$x_by/5,
  y_start = vf$y_start,
  y_end = vf$y_end,
  y_by = vf$y_by/5,
  x_sparse = 5,
  y_sparse = 5,
  ...
)
}
\arguments{
\item{vf}{A \code{vectorfield} object estimated by \code{\link[=fit_vf_2d]{fit_vf_2d()}}. The other parameters will be inherited from the \code{vectorfield} object if not otherwise specified.}

\item{x_start, x_end, x_by, y_start, y_end, y_by}{Specify the sample grid for the output: the starting, ending, and increment of the sequences for both axes.}

\item{x_sparse, y_sparse}{A number. For calculating the non-gradient part of the vector field, how much should the sample points be sparser than the sample points for the potential landscape?}

\item{...}{Other parameters passed to \code{\link[=approxPot2D]{approxPot2D()}}.}
}
\description{
This is done by a decomposition method described in the reference, using a modified version of \code{\link[waydown:approxPot2D]{waydown::approxPot2D()}}.
}
\references{
Rodríguez-Sánchez, P., Nes, E. H. van, & Scheffer, M. (2020). Climbing Escher’s stairs: A way to approximate stability landscapes in multidimensional systems. PLOS Computational Biology, 16(4), e1007788. https://doi.org/10.1371/journal.pcbi.1007788
}
\seealso{
\code{\link[=plot.2d_vf_landscape]{plot.2d_vf_landscape()}}
}
