% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscape_1d.R
\name{make_1d_ld}
\alias{make_1d_ld}
\title{Create a 1D potential landscape from a 1D vector field}
\usage{
make_1d_ld(vf, linear_interp = TRUE, n_grid = 200L)
}
\arguments{
\item{vf}{A \verb{1d_vectorfield} or \code{cv_vectorfield} object.}

\item{linear_interp}{Logical indicating whether to use interpolation in
predictions.}

\item{n_grid}{Number of grid points.}
}
\value{
A \verb{1d_static_ld} object.
}
\description{
This is the 1D analogue of \code{\link[=make_2d_ld]{make_2d_ld()}}, using the drift and diffusion
estimates from a fitted 1D vector field.
}
