% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vf_ld_gen2.R
\name{make_2d_ld}
\alias{make_2d_ld}
\title{Creates a 2D Potential Landscape object from a vector field.}
\usage{
make_2d_ld(vf, linear_interp = TRUE, n_grid = 100)
}
\arguments{
\item{vf}{A \code{vectorfield} or \code{cv_vectorfield} object representing the vector field.}

\item{linear_interp}{Logical indicating whether to use linear interpolation in predictions.}

\item{n_grid}{The number of grid points along one dimension (e.g., 100).}
}
\value{
An object of class \verb{2d_static_ld} containing:
- \code{dist}: A data frame with columns x, y, d (steady-state distribution), U (potential).
- \code{plot}: A plotly surface plot of the potential landscape U.
- \code{plot_2}: A ggplot2 raster plot of the potential landscape U.
- \code{vf}: The input vector field object.
- \code{ss}: The steady-state distribution matrix.
}
\description{
Creates a 2D Potential Landscape object from a vector field.
}
