% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseVFC.R
\name{norm2}
\alias{norm2}
\title{Normalize two matrices separately}
\usage{
norm2(x, y)
}
\arguments{
\item{x, y}{The two matrices.}
}
\description{
Normalize two matrices separately
}
\seealso{
\code{\link[=SparseVFC]{SparseVFC()}}
}
