% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_vectorfield.R
\name{normalize_predict_f}
\alias{normalize_predict_f}
\title{Return a normalized prediction function}
\usage{
normalize_predict_f(vf)
}
\arguments{
\item{vf}{A \code{vectorfield} object estimated by \code{\link[=fit_2d_vf]{fit_2d_vf()}}.}
}
\value{
A function that takes a vector \code{x} and returns a list of \code{v}, the drift part, and \code{a}, the diffusion part.
}
\description{
Return a normalized prediction function
}
