% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_api.R
\name{plot.2d_static_ld}
\alias{plot.2d_static_ld}
\alias{plot.1d_static_ld}
\alias{plot.1d_ld}
\alias{plot.2d_ld}
\alias{plot.2d_MVKE_landscape}
\alias{plot.2d_pf}
\title{Deprecated landscape and probability-flow plot methods}
\usage{
\method{plot}{`2d_static_ld`}(x, index = 1, ...)

\method{plot}{`1d_static_ld`}(x, index = 1, ...)

\method{plot}{`1d_ld`}(x, index = 1, ...)

\method{plot}{`2d_ld`}(x, index = 1, ...)

\method{plot}{`2d_MVKE_landscape`}(x, index = 1, ...)

\method{plot}{`2d_pf`}(x, ...)
}
\arguments{
\item{x}{A fitlandr object.}

\item{index}{The legacy landscape plot index.}

\item{...}{Arguments passed to the replacement method.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
Use \code{\link[simlandr:plotly_ld]{plotly_ld()}} for interactive three-dimensional landscapes and
\code{\link[ggplot2:autoplot]{autoplot()}} for ggplot output.
}
