% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot.2d_vf_landscape}
\alias{plot.2d_vf_landscape}
\title{Plot a 2D landscape estimated from a vector field.}
\usage{
\method{plot}{`2d_vf_landscape`}(
  x,
  type = c("landscape", "error"),
  show_vectorfield = FALSE,
  show_nongrad_vectorfield = FALSE,
  nongrad_vectorfield_enlarge = 1,
  nongrad_vectorfield_options = list(color = "skyblue"),
  arrow = grid::arrow(length = grid::unit(0.1, "cm")),
  ...
)
}
\arguments{
\item{x}{A \verb{2d_vf_landscape} object created by \code{\link[=fit_vfld_2d]{fit_vfld_2d()}}.}

\item{type}{One of "landscape" or "error"}

\item{show_vectorfield}{Should the vector field also be plotted? If so, the landscape will be plotted below the vector field. \code{FALSE} by default.}

\item{show_nongrad_vectorfield}{Should the non-gradient part of the vector field be plotted? If so, it will be plotted at the top layer. \code{FALSE} by default.}

\item{nongrad_vectorfield_enlarge}{A number. How many times should the non-gradient vectors be enlarged on the plot?}

\item{nongrad_vectorfield_options}{A list passing other customized parameters to \code{\link[ggplot2:geom_segment]{ggplot2::geom_segment()}} to control the vectors representing the non-gradient part of the vector field. Sky blue by default.}

\item{arrow}{The description of the arrow heads of the vectors on the plot (representing the vector field). Generated by \code{\link[grid:arrow]{grid::arrow()}}. Also see the \code{arrow} parameter of \code{\link[ggplot2:geom_segment]{ggplot2::geom_segment()}}.}

\item{...}{Other parameters passed to \code{\link[=plot.vectorfield]{plot.vectorfield()}}.}
}
\description{
Plot a 2D landscape estimated by \code{\link[=fit_vfld_2d]{fit_vfld_2d()}}. Powered by \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}.
}
