% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waydown_modified.R
\name{vectorfield_nongradient_2D}
\alias{vectorfield_nongradient_2D}
\title{Calculate the nongradient part of the vector field.}
\usage{
vectorfield_nongradient_2D(wdresult, x_sparse, y_sparse, ...)
}
\arguments{
\item{wdresult}{Output from \code{approxPot2D()}.}

\item{x_sparse, y_sparse}{Should the sample points for calculating the non-gradient part of the field be sparser than the sample points for calculating the landscape? Integer numbers larger than 1.}
}
\description{
Calculate the nongradient part of the vector field.
}
\author{
Jingmeng Cui
}
\keyword{internal}
