% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorfield_landscape.R
\name{VFCf}
\alias{VFCf}
\title{Generate a function to calculate the estimated vector field}
\usage{
VFCf(VFCresult)
}
\arguments{
\item{VFCresult}{A VFC result estimated from \code{\link[SparseVFC:SparseVFC]{SparseVFC::SparseVFC()}}.}
}
\description{
This will generate a helper function to estimate the vector field
for any position, based on the VFC estimation.
}
