% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_api.R
\name{autoplot.2d_static_ld}
\alias{autoplot.2d_static_ld}
\alias{autoplot.1d_static_ld}
\alias{autoplot.1d_ld}
\alias{autoplot.2d_ld}
\alias{autoplot.2d_MVKE_landscape}
\alias{autoplot.2d_pf}
\title{Autoplot a landscape or probability flow}
\usage{
\method{autoplot}{`2d_static_ld`}(object, ...)

\method{autoplot}{`1d_static_ld`}(object, ...)

\method{autoplot}{`1d_ld`}(object, ...)

\method{autoplot}{`2d_ld`}(object, ...)

\method{autoplot}{`2d_MVKE_landscape`}(object, ...)

\method{autoplot}{`2d_pf`}(object, ...)
}
\arguments{
\item{object}{A fitlandr object.}

\item{...}{Additional arguments, currently unused.}
}
\value{
A ggplot object.
}
\description{
These methods return the two-dimensional ggplot representation of a
landscape or probability-flow object.
}
