% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_generation.R
\name{bootstrap_2d_ld}
\alias{bootstrap_2d_ld}
\title{Generates bootstrap landscapes from bootstrap vector fields.}
\usage{
bootstrap_2d_ld(boot_vf, ...)
}
\arguments{
\item{boot_vf}{A \code{bootstrap_2d_vf} object containing bootstrap vector fields.}

\item{...}{Additional arguments passed to \code{make_2d_ld}.}
}
\value{
An object of class \code{bootstrap_2d_ld} containing:
- \code{bootstrap_lds}: A list of 2D landscapes from each bootstrap vector field. Note that the plots are removed to save space.
- \code{original_ld}: The original 2D landscape from the original vector field.
- \code{n_boot}: The number of bootstrap samples.
}
\description{
Generates bootstrap landscapes from bootstrap vector fields.
}
