% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vf_ld_gen2.R
\name{find_loc_min}
\alias{find_loc_min}
\title{Finds local minima in a 2D landscape object.}
\usage{
find_loc_min(ld, exclude_minor = TRUE, min_barrier = 0.05)
}
\arguments{
\item{ld}{A \verb{2d_ld} or \verb{2d_static_ld} object representing the landscape.}

\item{exclude_minor}{Logical indicating whether to mark minor local minima
based on the barrier height criterion, so that they can be easily excluded
from subsequent calculations. Default is TRUE.}

\item{min_barrier}{When the barrier height between a local minimum and any of the
saddle points connecting it to other local minima is less than \code{min_barrier} times
the highest barrier height value, the local minimum will be considered minor.
Default is 0.1.}
}
\value{
A data frame with columns x, y, U for each local minimum found.
}
\description{
Finds local minima in a 2D landscape object.
}
