% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorfield_landscape.R
\name{fit_vf_2d}
\alias{fit_vf_2d}
\title{Estimate a 2D vector field}
\usage{
fit_vf_2d(
  d,
  x,
  y,
  x_start,
  x_end,
  x_by,
  y_start,
  y_end,
  y_by,
  vector_position = "middle",
  na_action = "omit_data_points",
  ...
)
}
\arguments{
\item{d}{The data set used for estimating the vector field.
Should be a data frame or a matrix.}

\item{x, y}{Characters to indicate the name of the two variables.}

\item{vector_position}{One of \code{"start"}, \code{"middle"}, or \code{"end"}.}

\item{na_action}{One of \code{"omit_data_points"} or \code{"omit_vectors"}. If using \code{"omit_data_points"}, then the vectors will be retained even if there are \code{NA} points between them. If using \code{"omit_vectors"}, then the vectors will be omitted if either of its points is \code{NA}.}

\item{...}{Other parameters to be passed to \code{\link[SparseVFC:SparseVFC]{SparseVFC::SparseVFC()}}.}
}
\value{
A \code{vectorfield} object.
}
\description{
Estimate a 2D vector field
}
