% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorfield_landscape.R
\name{fit_vf_2d}
\alias{fit_vf_2d}
\title{Estimate a 2D vector field}
\usage{
fit_vf_2d(
  d,
  x,
  y,
  vector_position = "middle",
  x_start,
  x_end,
  x_by,
  y_start,
  y_end,
  y_by,
  ...
)
}
\arguments{
\item{d}{The data set used for estimating the vector field.
Should be a data frame or a matrix.}

\item{x, y}{Characters to indicate the name of the two variables.}

\item{vector_position}{One of "start", "middle", or "end".}

\item{...}{Other parameters to be passed to \link{SparseVFC}.}
}
\value{
A \code{vectorfield} object.
}
\description{
Estimate a 2D vector field
}
