% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitlandr-package.R
\docType{package}
\name{fitlandr-package}
\alias{fitlandr}
\alias{fitlandr-package}
\title{fitlandr: Fit Vector Fields and Potential Landscapes from Intensive Longitudinal Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A toolbox for estimating vector fields from intensive longitudinal data, and construct potential landscapes thereafter. The vector fields can be estimated with two nonparametric methods: the Multivariate Vector Field Kernel Estimator (MVKE) by Bandi & Moloche (2018) \doi{10.1017/S0266466617000305} and the Sparse Vector Field Consensus (SparseVFC) algorithm by Ma et al. (2013) \doi{10.1016/j.patcog.2013.05.017}. The potential landscapes can be constructed with a simulation-based approach with the simlandr package (Cui et al., 2021) \doi{10.31234/osf.io/pzva3}, or the Bhattacharya et al. (2011) method for path integration \doi{10.1186/1752-0509-5-85}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://sciurus365.github.io/fitlandr/}
  \item \url{https://github.com/Sciurus365/fitlandr}
  \item Report bugs at \url{https://github.com/Sciurus365/fitlandr/issues}
}

}
\author{
\strong{Maintainer}: Jingmeng Cui \email{jingmeng.cui@outlook.com} (\href{https://orcid.org/0000-0003-3421-8457}{ORCID})

}
\keyword{internal}
