% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitlandr-package.R
\docType{package}
\name{fitlandr-package}
\alias{fitlandr}
\alias{fitlandr-package}
\title{fitlandr: Fit a Potential Landscape from Data}
\description{
Fit a potential function from data.
}
\author{
\strong{Maintainer}: Jingmeng Cui \email{jingmeng.cui@outlook.com} (\href{https://orcid.org/0000-0003-3421-8457}{ORCID})

}
\keyword{internal}
