% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernal_estimator.R
\name{MVKE}
\alias{MVKE}
\title{Multivariate vector field kernel estimator}
\usage{
MVKE(d, h = 1)
}
\arguments{
\item{d}{The dataset. Should be a matrix or a data frame, with each row representing a random vector.}

\item{h}{The bandwidth for the kernel estimator.}
}
\value{
A function(x), which then returns the \eqn{\mu} and \eqn{a} estimators at the position \eqn{x}.
}
\description{
Gaussian product kernel is used.
}
\references{
Bandi, F. M., & Moloche, G. (2018). On the functional estimation of multivariate diffusion processes. Econometric Theory, 34(4), 896-946. https://doi.org/10.1017/S0266466617000305
}
