% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot.1d_vectorfield}
\alias{plot.1d_vectorfield}
\title{Plot a 2D vector field}
\usage{
\method{plot}{`1d_vectorfield`}(x, ...)
}
\arguments{
\item{x}{A \code{vectorfield} object estimated by \code{\link[=fit_2d_vf]{fit_2d_vf()}}.}

\item{...}{Arguments passed to \code{\link[=autoplot]{autoplot()}}.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
\code{plot.vectorfield()} is deprecated in favor of \code{\link[=autoplot]{autoplot()}}.
}
