% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot.2d_vf_landscape}
\alias{plot.2d_vf_landscape}
\title{Plot a 2D landscape estimated from a vector field.}
\usage{
\method{plot}{`2d_vf_landscape`}(x, vectorfield = FALSE, type = c("landscape", "error"), ...)
}
\arguments{
\item{x}{A \verb{2d_vf_landscape} object created by \code{\link[=fit_2d_vfld]{fit_2d_vfld()}}.}

\item{vectorfield}{Should the vector field also be plotted? If so, the landscape will be ploted below the vector field. \code{FALSE} by default.}

\item{type}{One of "landscape" or "error".}

\item{...}{Other parameters passed to \code{\link[=plot.vectorfield]{plot.vectorfield()}}.}
}
\description{
Plot a 2D landscape estimated by \code{\link[=fit_2d_vfld]{fit_2d_vfld()}}. Powered by \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}.
}
