% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapping.R
\name{plot.summary_bootstrap_2d_ld}
\alias{plot.summary_bootstrap_2d_ld}
\title{Plots all the bootstrap minima positions}
\usage{
\method{plot}{summary_bootstrap_2d_ld}(x, ...)
}
\arguments{
\item{x}{A \code{summary_bootstrap_2d_ld} object.}

\item{...}{Additional arguments (not used).}
}
\description{
Color represents the U value at that minimum.
}
