% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-functions.R
\name{reorder_output}
\alias{reorder_output}
\title{Reorder a simulation output in time order}
\usage{
reorder_output(s, chains)
}
\arguments{
\item{s}{A simulation output, possibly generated by \code{\link[=sim_vf]{sim_vf()}}}

\item{chains}{How many chains simulations should be performed?}
}
\value{
A reordered matrix of the simulation output.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function was deprecated in favor of the more efficient and accurate \code{\link[=make_2d_ld]{make_2d_ld()}} approach.

Then \code{\link[simlandr:check_conv]{simlandr::check_conv()}} can be used meaningfully.
}
\keyword{internal}
