% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorfield_simulation.R
\name{reorder_output}
\alias{reorder_output}
\title{Reorder a simulation output in time order}
\usage{
reorder_output(s, chains)
}
\arguments{
\item{s}{A simulation output, possibly generated by \code{\link[=sim_vf]{sim_vf()}}}

\item{chains}{How many chains simulations should be performed?}
}
\value{
A reordered matrix of the simulation output.
}
\description{
Then \code{\link[simlandr:check_conv]{simlandr::check_conv()}} can be used meaningfully.
}
