% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorfield_simulation.R
\name{sim_vf_options}
\alias{sim_vf_options}
\title{Options controlling the vector field simulation}
\usage{
sim_vf_options(
  vf,
  chains = 10,
  length = 10000,
  discard = 0.3,
  inits = matrix(c(runif(chains, min = vf$x_start, max = vf$x_end), runif(chains, min =
    vf$y_start, max = vf$y_end)), ncol = 2)
)
}
\arguments{
\item{vf}{A \code{vectorfield} object estimated by \code{\link[=fit_vf_2d]{fit_vf_2d()}}.]}

\item{chains}{How many chains simulations should be performed?}

\item{length}{The simulation length for each chain.}

\item{discard}{How much of the starting part of each chain should be discarded? (Warming-up period.)}

\item{inits}{The initial values of each chain.}
}
\description{
Options controlling the vector field simulation
}
