% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorfield_simulation.R
\name{sim_vf_options}
\alias{sim_vf_options}
\title{Options controlling the vector field simulation
See \code{\link[=sim_vf]{sim_vf()}} for details.}
\usage{
sim_vf_options(
  vf,
  noise = 1,
  chains = 10,
  length = 10000,
  discard = 0.3,
  forbid_overflow = FALSE,
  inits = rlang::expr(matrix(c(stats::runif(chains, min = vf$lims[1], max = vf$lims[2]),
    stats::runif(chains, min = vf$lims[3], max = vf$lims[4])), ncol = 2))
)
}
\arguments{
\item{vf}{A \code{vectorfield} object estimated by \code{\link[=fit_2d_vf]{fit_2d_vf()}}.}

\item{noise}{Relative noise of the simulation. Set this smaller when the simulation is unstable (e.g., when the elements in the diffusion matrix are not finite), and set this larger when the simulation converges too slowly.}

\item{chains}{How many chains simulations should be performed?}

\item{length}{The simulation length for each chain.}

\item{discard}{How much of the starting part of each chain should be discarded? (Warming-up period.)}

\item{forbid_overflow}{If \code{TRUE}, when the simulated system runs out of the margins specified in \code{vf}, the system will be moved back to a random position within the margins. This can help to stablize the simulation. \code{FALSE} by default.}

\item{inits}{The initial values of each chain.}
}
\description{
Options controlling the vector field simulation
See \code{\link[=sim_vf]{sim_vf()}} for details.
}
