% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorfield_simulation.R
\name{simlandr_options}
\alias{simlandr_options}
\title{Options controlling the landscape construction}
\usage{
simlandr_options(
  vf,
  x = rlang::expr(vf$x),
  y = rlang::expr(vf$y),
  n = 200,
  lims = rlang::expr(c(vf$x_start, vf$x_end, vf$y_start, vf$y_end)),
  Umax = 8,
  h = 0.1,
  kde_fun = "ks"
)
}
\arguments{
\item{vf}{A \code{vectorfield} object estimated by \code{\link[=fit_2d_vf]{fit_2d_vf()}}.]}

\item{x, y}{The name of the target variable.}

\item{n, lims, h}{Passed to \code{\link[ks:kde]{ks::kde()}} or \code{\link[MASS:kde2d]{MASS::kde2d()}}.
If using \code{ks::kde}, \code{H = diag(h, 2, 2)}.
Note: the definition of bandwidth (\code{h}) is different in two functions.
To get a similar output, the \code{h} is about 50 to 5000 times smaller for \code{\link[ks:kde]{ks::kde()}} than \code{\link[MASS:kde2d]{MASS::kde2d()}}}

\item{Umax}{The maximum displayed value of potential.}

\item{kde_fun}{Which to use? Choices: "ks" \code{ks::kde} (default; faster and taking less memory); "MASS" \code{MASS::kde2d}.}
}
\description{
Options controlling the landscape construction
}
