% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseVFC.R
\name{SparseVFC}
\alias{SparseVFC}
\title{Sparse Vector Field Consensus}
\usage{
SparseVFC(
  X,
  Y,
  MaxIter = 500,
  gamma = 0.9,
  beta = 0.1,
  lambda = 3,
  theta = 0.75,
  a = 10,
  ecr = 1e-05,
  minP = 1e-05,
  silent = TRUE
)
}
\description{
Translated from codes in https://github.com/jiayi-ma/VFC
Jiayi Ma, Ji Zhao, Jinwen Tian, Xiang Bai, and Zhuowen Tu. "Regularized Vector Field Learning with Sparse Approximation for Mismatch Removal", Pattern Recognition, 46(12), pp. 3519-3532, 2013.
}
