% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vf_ld_gen2.R
\name{ss_fp_2d}
\alias{ss_fp_2d}
\title{Estimates the steady-state distribution using Central Differencing for an
arbitrary rectangular domain defined by x_range and y_range.}
\usage{
ss_fp_2d(vf, linear_interp = TRUE, n_grid = 100)
}
\arguments{
\item{vf}{A \code{vectorfield} or \code{cv_vectorfield} object representing the vector field.}

\item{linear_interp}{Logical indicating whether to use linear interpolation in predictions.}

\item{n_grid}{The number of grid points along one dimension (e.g., 100).}
}
\value{
matrix of probability density.
}
\description{
Estimates the steady-state distribution using Central Differencing for an
arbitrary rectangular domain defined by x_range and y_range.
}
